+++ mirrord/config/src/lib.rs
@@ -227,11 +227,11 @@ pub struct LayerConfig {
 
     /// ## operator {#root-operator}
     ///
-    /// Allow to lookup if operator is installed on cluster and use it.
-    ///
-    /// Defaults to `true`.
-    #[config(env = "MIRRORD_OPERATOR_ENABLE", default = true)]
-    pub operator: bool,
+    /// Whether mirrord should use the operator.
+    /// If not set, mirrord will first attempt to use the operator, but continue without it in case
+    /// of failure.
+    #[config(env = "MIRRORD_OPERATOR_ENABLE")]
+    pub operator: Option<bool>,
 
     /// ## kubeconfig {#root-kubeconfig}
     ///
@@ -434,7 +434,7 @@ impl LayerConfig {
         }
 
         if self.feature.copy_target.enabled {
-            if !self.operator {
+            if self.operator == Some(false) {
                 return Err(ConfigError::Conflict(
                     "The copy target feature requires a mirrord operator, \
                    please either disable this option or use the operator."
@@ -442,7 +442,8 @@ impl LayerConfig {
                 ));
             }
 
-            if self.target.path.is_none() {
+            // Target may also be set later in the UI.
+            if self.target.path.is_none() && !context.ide {
                 return Err(ConfigError::Conflict(
                     "The copy target feature is not compatible with a targetless agent, \
                     please either disable this option or specify a target."