+++ mirrord/operator/src/client.rs
@@ -3,7 +3,7 @@ use std::io;
 use base64::{engine::general_purpose, Engine as _};
 use chrono::{DateTime, NaiveDate, NaiveTime, Utc};
 use futures::{SinkExt, StreamExt};
-use http::{request::Request, StatusCode};
+use http::request::Request;
 use kube::{api::PostParams, Api, Client, Resource};
 use mirrord_analytics::{AnalyticsHash, AnalyticsOperatorProperties, AnalyticsReporter};
 use mirrord_auth::{
@@ -176,9 +176,6 @@ impl OperatorApi {
     }
 
     /// Creates new [`OperatorSessionConnection`] based on the given [`LayerConfig`].
-    ///
-    /// Returns [`None`] if the operator is not found.
-    ///
     /// Keep in mind that some failures here won't stop mirrord from hooking into the process
     /// and working, it'll just work without the operator.
     ///
@@ -188,16 +185,13 @@ impl OperatorApi {
         config: &LayerConfig,
         progress: &P,
         analytics: &mut AnalyticsReporter,
-    ) -> Result<Option<OperatorSessionConnection>>
+    ) -> Result<OperatorSessionConnection>
     where
         P: Progress + Send + Sync,
     {
         let operator_api = OperatorApi::new(config).await?;
 
-        let Some(operator) = operator_api.fetch_operator().await? else {
-            // No operator found.
-            return Ok(None);
-        };
+        let operator = operator_api.fetch_operator().await?;
 
         // Warns the user if their license is close to expiring.
         //
@@ -317,7 +311,7 @@ impl OperatorApi {
         };
         let connection = operator_api.connect_target(session_info).await?;
 
-        Ok(Some(connection))
+        Ok(connection)
     }
 
     /// Connects to exisiting operator session based on the given [`LayerConfig`] and
@@ -372,16 +366,14 @@ impl OperatorApi {
     }
 
     #[tracing::instrument(level = "debug", skip(self), ret)]
-    async fn fetch_operator(&self) -> Result<Option<MirrordOperatorCrd>> {
+    async fn fetch_operator(&self) -> Result<MirrordOperatorCrd> {
         let api: Api<MirrordOperatorCrd> = Api::all(self.client.clone());
-        match api.get(OPERATOR_STATUS_NAME).await {
-            Ok(crd) => Ok(Some(crd)),
-            Err(kube::Error::Api(e)) if e.code == StatusCode::NOT_FOUND => Ok(None),
-            Err(error) => Err(OperatorApiError::KubeError {
+        api.get(OPERATOR_STATUS_NAME)
+            .await
+            .map_err(|error| OperatorApiError::KubeError {
                 error,
                 operation: "finding operator in the cluster".into(),
-            }),
-        }
+            })
     }
 
     /// See `operator/controller/src/target.rs::TargetProvider::get_resource`.